REST API doesn't start; generating stager fails with REST API call
The REST API won’t start due to a bug:

# ./empire --rest --username "emp" --password "emp"

[*] Loading modules from: /mnt/hgfs/cjones/Empire/lib/modules/
 * Starting Empire RESTful API on port: 1337
 * RESTful API token: 2bjmeuwa6pr6yy4x0n88rauyyl1nve7cekdgkefh
Traceback (most recent call last):
  File "/usr/lib/python2.7/logging/__init__.py", line 853, in emit
    msg = self.format(record)
  File "/usr/lib/python2.7/logging/__init__.py", line 726, in format
    return fmt.format(record)
  File "/usr/lib/python2.7/logging/__init__.py", line 465, in format
    record.message = record.getMessage()
  File "/usr/lib/python2.7/logging/__init__.py", line 329, in getMessage
    msg = msg % self.args
TypeError: %d format: a number is required, not str
Logged from file _internal.py, line 87

After casting the port from a string to an int, the REST service works.

Also, when calling the API to create a stager, the .dll is not actually
written to disk.