PowerUp.ps1:883 - Remove "C:\" False Positives
By default, the "C:\" folder has special permissions set that allow unprivileged users to create new folders, but not to create or modify files in C:\.  This special premission set is misinterpreted by the script as a writable folder in the path, even though it is not.  The proposed change removes results that return "C:\" (subfolders within C:\ are still returned if they are modifiable).