stderr + ruby fiddle support
This updates the ruby code to use fiddle library for ruby 2.3 and above while still supporting 'DL' for older versions. 'DL' is usually not found on new ruby installs.

Additionally if any errors occur when it tries to run or while it's running, they will be logged to the keylog file instead of outputed to the terminal.